Properly split connect and TLS features
